backend/requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.12
Pillow>=8.0
django-cors-headers>=3.7
django-environ>=0.8
drf-yasg>=1.21
